requirements.txt
# Core RAG Dependencies
chromadb>=0.4.22
pymupdf>=1.23.0
sentence-transformers>=2.2.2

# Model Providers
mistralai>=0.1.0
ollama>=0.1.0

# Search and Ranking
rank-bm25>=0.2.2

# Configuration and Environment
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
tiktoken>=0.5.0

# Evaluation Metrics
scikit-learn>=1.3.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
